Added timeout and disable ssl verify options, fixed issue with http basic auth
Some requests from datasets could take longer than the default 10 second timeout that requests library uses, so an option to override the timeout has been added with a new default of 30 seconds rather than 10.

In some instances it may be necessary to skip SSL/TLS verification if the target server has a wonky certificate.  This is obviously not a secure approach and should only be used if you know and trust the target server - SSL/TLS verification is enabled by default.

Pandas read_csv() function doesn't handle http basic auth URLs well, so that has been abstracted into a request call and that is then passed to pandas to resolve the authentication issue.